Updating interleaves for Neon and AVX512
