+++ field/src/interleaves.rs
@@ -1,12 +1,52 @@
 //! A file containing a collection of architecture-specific interleaving functions.
 //! Used for PackedFields to implement interleaving operations.
 
+#[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
+pub mod interleave {
+    use core::arch::aarch64::{self, uint32x4_t};
+
+    #[inline]
+    #[must_use]
+    pub fn interleave_u32(v0: uint32x4_t, v1: uint32x4_t) -> (uint32x4_t, uint32x4_t) {
+        // We want this to compile to:
+        //      trn1  res0.4s, v0.4s, v1.4s
+        //      trn2  res1.4s, v0.4s, v1.4s
+        // throughput: .5 cyc/2 vec (16 els/cyc)
+        // latency: 2 cyc
+        unsafe {
+            // Safety: If this code got compiled then NEON intrinsics are available.
+            (aarch64::vtrn1q_u32(v0, v1), aarch64::vtrn2q_u32(v0, v1))
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn interleave_u64(v0: uint32x4_t, v1: uint32x4_t) -> (uint32x4_t, uint32x4_t) {
+        // We want this to compile to:
+        //      trn1  res0.2d, v0.2d, v1.2d
+        //      trn2  res1.2d, v0.2d, v1.2d
+        // throughput: .5 cyc/2 vec (16 els/cyc)
+        // latency: 2 cyc
+
+        // To transpose 64-bit blocks, cast the [u32; 4] vectors to [u64; 2], transpose, and cast back.
+        unsafe {
+            // Safety: If this code got compiled then NEON intrinsics are available.
+            let v0 = aarch64::vreinterpretq_u64_u32(v0);
+            let v1 = aarch64::vreinterpretq_u64_u32(v1);
+            (
+                aarch64::vreinterpretq_u32_u64(aarch64::vtrn1q_u64(v0, v1)),
+                aarch64::vreinterpretq_u32_u64(aarch64::vtrn2q_u64(v0, v1)),
+            )
+        }
+    }
+}
+
 #[cfg(all(
     target_arch = "x86_64",
     target_feature = "avx2",
     not(all(feature = "nightly-features", target_feature = "avx512f"))
 ))]
-pub mod interleave_avx2 {
+pub mod interleave {
     use core::arch::x86_64::{self, __m256i};
 
     #[inline]
@@ -88,6 +128,193 @@ pub mod interleave_avx2 {
             (
                 x86_64::_mm256_blend_epi32::<0b11110000>(a, t),
                 x86_64::_mm256_blend_epi32::<0b11110000>(t, b),
+            )
+        }
+    }
+}
+
+#[cfg(all(
+    feature = "nightly-features",
+    target_arch = "x86_64",
+    target_feature = "avx512f"
+))]
+pub mod interleave {
+    use core::arch::x86_64::{self, __m512i, __mmask8, __mmask16};
+
+    const EVENS: __mmask16 = 0b0101010101010101;
+    const EVENS4: __mmask16 = 0x0f0f;
+
+    // vpshrdq requires AVX-512VBMI2.
+    #[cfg(target_feature = "avx512vbmi2")]
+    #[inline]
+    #[must_use]
+    fn interleave1_antidiagonal(x: __m512i, y: __m512i) -> __m512i {
+        unsafe {
+            // Safety: If this code got compiled then AVX-512VBMI2 intrinsics are available.
+            x86_64::_mm512_shrdi_epi64::<32>(x, y)
+        }
+    }
+
+    // If we can't use vpshrdq, then do a vpermi2d, but we waste a register and double the latency.
+    #[cfg(not(target_feature = "avx512vbmi2"))]
+    #[inline]
+    #[must_use]
+    fn interleave1_antidiagonal(x: __m512i, y: __m512i) -> __m512i {
+        const INTERLEAVE1_INDICES: __m512i = unsafe {
+            // Safety: `[u32; 16]` is trivially transmutable to `__m512i`.
+            transmute::<[u32; WIDTH], _>([
+                0x01, 0x10, 0x03, 0x12, 0x05, 0x14, 0x07, 0x16, 0x09, 0x18, 0x0b, 0x1a, 0x0d, 0x1c,
+                0x0f, 0x1e,
+            ])
+        };
+        unsafe {
+            // Safety: If this code got compiled then AVX-512F intrinsics are available.
+            x86_64::_mm512_permutex2var_epi32(x, INTERLEAVE1_INDICES, y)
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn interleave_u32(x: __m512i, y: __m512i) -> (__m512i, __m512i) {
+        // If we have AVX-512VBMI2, we want this to compile to:
+        //      vpshrdq    t, x, y, 32
+        //      vpblendmd  res0 {EVENS}, t, x
+        //      vpblendmd  res1 {EVENS}, y, t
+        // throughput: 1.5 cyc/2 vec (21.33 els/cyc)
+        // latency: 2 cyc
+        //
+        // Otherwise, we want it to compile to:
+        //      vmovdqa32  t, INTERLEAVE1_INDICES
+        //      vpermi2d   t, x, y
+        //      vpblendmd  res0 {EVENS}, t, x
+        //      vpblendmd  res1 {EVENS}, y, t
+        // throughput: 1.5 cyc/2 vec (21.33 els/cyc)
+        // latency: 4 cyc
+
+        // We currently have:
+        //   x = [ x0  x1  x2  x3  x4  x5  x6  x7  x8  x9  xa  xb  xc  xd  xe  xf ],
+        //   y = [ y0  y1  y2  y3  y4  y5  y6  y7  y8  y9  ya  yb  yc  yd  ye  yf ].
+        // First form
+        //   t = [ x1  y0  x3  y2  x5  y4  x7  y6  x9  y8  xb  ya  xd  yc  xf  ye ].
+        let t = interleave1_antidiagonal(x, y);
+
+        unsafe {
+            // Safety: If this code got compiled then AVX-512F intrinsics are available.
+
+            // Then
+            //   res0 = [ x0  y0  x2  y2  x4  y4  x6  y6  x8  y8  xa  ya  xc  yc  xe  ye ],
+            //   res1 = [ x1  y1  x3  y3  x5  y5  x7  y7  x9  y9  xb  yb  xd  yd  xf  yf ].
+            (
+                x86_64::_mm512_mask_blend_epi32(EVENS, t, x),
+                x86_64::_mm512_mask_blend_epi32(EVENS, y, t),
+            )
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    fn shuffle_epi64<const MASK: i32>(a: __m512i, b: __m512i) -> __m512i {
+        // The instruction is only available in the floating-point flavor; this distinction is only for
+        // historical reasons and no longer matters. We cast to floats, do the thing, and cast back.
+        unsafe {
+            let a = x86_64::_mm512_castsi512_pd(a);
+            let b = x86_64::_mm512_castsi512_pd(b);
+            x86_64::_mm512_castpd_si512(x86_64::_mm512_shuffle_pd::<MASK>(a, b))
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn interleave_u64(x: __m512i, y: __m512i) -> (__m512i, __m512i) {
+        // We want this to compile to:
+        //      vshufpd    t, x, y, 55h
+        //      vpblendmq  res0 {EVENS}, t, x
+        //      vpblendmq  res1 {EVENS}, y, t
+        // throughput: 1.5 cyc/2 vec (21.33 els/cyc)
+        // latency: 2 cyc
+
+        unsafe {
+            // Safety: If this code got compiled then AVX-512F intrinsics are available.
+
+            // We currently have:
+            //   x = [ x0  x1  x2  x3  x4  x5  x6  x7  x8  x9  xa  xb  xc  xd  xe  xf ],
+            //   y = [ y0  y1  y2  y3  y4  y5  y6  y7  y8  y9  ya  yb  yc  yd  ye  yf ].
+            // First form
+            //   t = [ x2  x3  y0  y1  x6  x7  y4  y5  xa  xb  y8  y9  xe  xf  yc  yd ].
+            let t = shuffle_epi64::<0b01010101>(x, y);
+
+            // Then
+            //   res0 = [ x0  x1  y0  y1  x4  x5  y4  y5  x8  x9  y8  y9  xc  xd  yc  yd ],
+            //   res1 = [ x2  x3  y2  y3  x6  x7  y6  y7  xa  xb  ya  yb  xe  xf  ye  yf ].
+            (
+                x86_64::_mm512_mask_blend_epi64(EVENS as __mmask8, t, x),
+                x86_64::_mm512_mask_blend_epi64(EVENS as __mmask8, y, t),
+            )
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn interleave_u128(x: __m512i, y: __m512i) -> (__m512i, __m512i) {
+        // We want this to compile to:
+        //      vmovdqa64   t, INTERLEAVE4_INDICES
+        //      vpermi2q    t, x, y
+        //      vpblendmd   res0 {EVENS4}, t, x
+        //      vpblendmd   res1 {EVENS4}, y, t
+        // throughput: 1.5 cyc/2 vec (21.33 els/cyc)
+        // latency: 4 cyc
+
+        const INTERLEAVE4_INDICES: __m512i = unsafe {
+            // Safety: `[u64; 8]` is trivially transmutable to `__m512i`.
+            transmute::<[u64; WIDTH / 2], _>([0o02, 0o03, 0o10, 0o11, 0o06, 0o07, 0o14, 0o15])
+        };
+
+        unsafe {
+            // Safety: If this code got compiled then AVX-512F intrinsics are available.
+
+            // We currently have:
+            //   x = [ x0  x1  x2  x3  x4  x5  x6  x7  x8  x9  xa  xb  xc  xd  xe  xf ],
+            //   y = [ y0  y1  y2  y3  y4  y5  y6  y7  y8  y9  ya  yb  yc  yd  ye  yf ].
+            // First form
+            //   t = [ x4  x5  x6  x7  y0  y1  y2  y3  xc  xd  xe  xf  y8  y9  ya  yb ].
+            let t = x86_64::_mm512_permutex2var_epi64(x, INTERLEAVE4_INDICES, y);
+
+            // Then
+            //   res0 = [ x0  x1  x2  x3  y0  y1  y2  y3  x8  x9  xa  xb  y8  y9  ya  yb ],
+            //   res1 = [ x4  x5  x6  x7  y4  y5  y6  y7  xc  xd  xe  xf  yc  yd  ye  yf ].
+            (
+                x86_64::_mm512_mask_blend_epi32(EVENS4, t, x),
+                x86_64::_mm512_mask_blend_epi32(EVENS4, y, t),
+            )
+        }
+    }
+
+    #[inline]
+    #[must_use]
+    pub fn interleave_u256(x: __m512i, y: __m512i) -> (__m512i, __m512i) {
+        // We want this to compile to:
+        //      vshufi64x2  t, x, b, 4eh
+        //      vpblendmq   res0 {EVENS4}, t, x
+        //      vpblendmq   res1 {EVENS4}, y, t
+        // throughput: 1.5 cyc/2 vec (21.33 els/cyc)
+        // latency: 4 cyc
+
+        unsafe {
+            // Safety: If this code got compiled then AVX-512F intrinsics are available.
+
+            // We currently have:
+            //   x = [ x0  x1  x2  x3  x4  x5  x6  x7  x8  x9  xa  xb  xc  xd  xe  xf ],
+            //   y = [ y0  y1  y2  y3  y4  y5  y6  y7  y8  y9  ya  yb  yc  yd  ye  yf ].
+            // First form
+            //   t = [ x8  x9  xa  xb  xc  xd  xe  xf  y0  y1  y2  y3  y4  y5  y6  y7 ].
+            let t = x86_64::_mm512_shuffle_i64x2::<0b01_00_11_10>(x, y);
+
+            // Then
+            //   res0 = [ x0  x1  x2  x3  x4  x5  x6  x7  y0  y1  y2  y3  y4  y5  y6  y7 ],
+            //   res1 = [ x8  x9  xa  xb  xc  xd  xe  xf  y8  y9  ya  yb  yc  yd  ye  yf ].
+            (
+                x86_64::_mm512_mask_blend_epi64(EVENS4 as __mmask8, t, x),
+                x86_64::_mm512_mask_blend_epi64(EVENS4 as __mmask8, y, t),
             )
         }
     }
