+++ goldilocks/src/x86_64_avx2/packing.rs
@@ -6,7 +6,7 @@ use core::mem::transmute;
 use core::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign};
 
 use p3_field::exponentiation::exp_10540996611094048183;
-use p3_field::interleave_avx2::{interleave_u64, interleave_u128};
+use p3_field::interleave::{interleave_u64, interleave_u128};
 use p3_field::op_assign_macros::{
     impl_add_assign, impl_add_base_field, impl_div_methods, impl_mul_base_field, impl_mul_methods,
     impl_rng, impl_sub_assign, impl_sub_base_field, impl_sum_prod_base_field, ring_sum,
